Refactored tsconfig.json Configuration
Removed the current config files and added two new config files:
tsconfig.json and tsconfig.build.json. One of the configuration files
compiles all sources and tests, the other one only compiles the sources.
One extends the other, but modifies the source and output directories.

The configuration used is copied from another personal project, and is
stricter than the original config. Fixed all new errors the compiler
flagged. All the methods that were returning undefined and the compiler
flagged as methods where not all paths returned a value are now
explicitely returning **null** instead. A future improvement would be to
chase all methods where return types are like **Type | undefined**, or
which are returning undefined explicitely. It is better to return null
in case a method decides there is no value to return.