Fix Majority eslint Errors Sources
Fix the majority of all the eslint errors in the sources. There are a
few rules ignored for now, with notes for next steps. Two test cases are
currently failing because of the change from **undefined** to **null**.
The rest of the test are still passing. Next step is to fix all eslint
errors in the test (and fix those two failing test cases).