Add Release Workflow and Scripts
Added a new github workflow to release the project on tags. There should
probably be check in there that tag version matches the package.json
version, but for now we just double check before pressing buttons.

This commit also introduces a few more scripts in the package.json file,
and a javascript module that trims the package.json that ships with the
release.