Move jsonc-parser to Dependencies
Move the jsonc-parser from the devDependencies to the normal
dependencies, as this module is used in the sources. This is a bug that
actually only appeared after the split of the repository from
vscode-psl, as vscode-psl has the jsonc-parser as normal dependency,
therefore the plugin never had issues resolving this dependency.