Move from old eslint modules to new ones
Move from the old @typscript-eslint modules and use the @eslint/js
instead. Also use the same rules as the profile-connector project. This
commit has a lot of changes on quotes, as the new rules force the use of
double quotes instead of single quotes, as this is the "standard" quotes
used in many languages.