Change unit test to use node built-in test
Dropped the use of jest and start to use the built-in test modules from
node instead. Together with the introduction of tsx it also removes the
need for two different tsconfig files.

This commit started fixing the tests, but there are still a few tests
failing that will be fixed in subsequent commits.