Update utilities to use fs instead of fs-extra
Update the utilities.ts file to use fs.readFile instead of
fs-extra.readFile. The unit test were failing with an error that should
actually not happen, but it was impossible to figure out why it was
failin the way it did (readFile not being defined on fs-extra, even
though it was in the default exports). Relying on built-in modules is
better anyway, therefore this is still considered a good alternative.

This commit also introduces vscode launch settings for debugging unit
tests using `node --test`, and removed all remaining jest related
config.